Filter mutations

This function filters the mutations based on the mutation calling model and parameters. The loci passed the filter will be saved in the h5 file, together with the filter parameters.

filter_loc(
  mtmutObj,
  min_cell = 5,
  model = "bb",
  p_threshold = 0.05,
  p_adj_method = "fdr"
)

Arguments

mtmutObj

a mtmutObj object.

min_cell

a integer of the minimum number of cells with mutation, the default is 5.

model

a string of the model for mutation calling, it can be "bb", "bm" or "bi" which stands for beta binomial, binomial mixture and binomial model respectively, the default is "bb".

p_threshold

a numeric of the p-value threshold, the default is 0.05.

p_adj_method

a string of the method for p-value adjustment, . refer to p.adjust. The default is "fdr".

Value

a mtmutObj object with loc_pass and loc_filter updated.

Examples

## Use the example data
f <- system.file("extdata", "mini_dataset.tsv.gz", package = "scMitoMut")

## Create a temporary h5 file
## In real case, we keep the h5 in project folder for future use
f_h5_tmp <- tempfile(fileext = ".h5")

## Load the data with parse_table function
f_h5 <- parse_table(f, sep = "\t", h5_file = f_h5_tmp)

## open the h5 file and create a mtmutObj object
x <- open_h5_file(f_h5)
run_model_fit(x)
#> chrM.200
#> chrM.204
#> chrM.310
#> chrM.824
#> chrM.1000
#> chrM.1001
#> chrM.1227
#> chrM.2285
#> chrM.6081
#> chrM.9429
#> chrM.9728
#> chrM.9804
#> chrM.9840
#> chrM.12889
#> chrM.16093
#> chrM.16147
#>           used (Mb) gc trigger  (Mb) limit (Mb) max used  (Mb)
#> Ncells 1476551 78.9    2382302 127.3         NA  2382302 127.3
#> Vcells 2904771 22.2    8388608  64.0      16384  8387717  64.0
x <- filter_loc(x, min_cell = 5, model = "bb", p_threshold = 0.05, p_adj_method = "fdr")
x
#> mtmutObj object
#> -------------------------------------------------
#> h5 file: /var/folders/46/vh_b6qzs5kzgdhvx1kcn7hyh0000gn/T//RtmpvCtNq4/file15552f81846a.h5
#> Available loci: 16
#> Selected loci: 16
#> Available cells: 1359
#> Selected cells: 1359
#> Loci passed the filter: 14
#> filter parameters: 
#> 	min_cell: 5
#> 	model: bb
#> 	p_threshold: 0.05
#> 	p_adj_method: fdr