Get p-value list for single locus

This function returns the p-value list for a single locus.

get_pval(mtmutObj, loc, model = "bb", method = "fdr")

Arguments

mtmutObj

a mtmutObj object.

loc

a string of the locus.

model

a string of the model for mutation calling, it can be "bb", "bm" or "bi" which stands for beta binomial, binomial mixture and binomial model respectively.

method

a string of the method for p-value adjustment, refer to p.adjust.

Value

a vector of p-value for each cell.

Examples

## Use the example data
f <- system.file("extdata", "mini_dataset.tsv.gz", package = "scMitoMut")
## Create a temporary h5 file
## In real case, we keep the h5 in project folder for future use
f_h5_tmp <- tempfile(fileext = ".h5")

## Load the data with parse_table function
f_h5 <- parse_table(f, sep = "\t", h5_file = f_h5_tmp)
## open the h5 file and create a mtmutObj object
x <- open_h5_file(f_h5)
run_model_fit(x)
#> chrM.200
#> chrM.204
#> chrM.310
#> chrM.824
#> chrM.1000
#> chrM.1001
#> chrM.1227
#> chrM.2285
#> chrM.6081
#> chrM.9429
#> chrM.9728
#> chrM.9804
#> chrM.9840
#> chrM.12889
#> chrM.16093
#> chrM.16147
#>           used (Mb) gc trigger  (Mb) limit (Mb) max used  (Mb)
#> Ncells 1476871 78.9    2382302 127.3         NA  2382302 127.3
#> Vcells 2900782 22.2    8388608  64.0      16384  8387717  64.0
get_pval(x, "chrM.1000", "bb", "fdr")
#>   [1] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>   [7] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [13] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [19] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [25] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [31] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [37] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [43] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [49] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [55] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [61] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [67] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [73] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [79] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [85] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [91] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [97] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [103] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [109] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [115] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [121] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [127] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [133] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [139] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [145] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [151] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [157] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [163] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [169] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [175] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [181] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [187] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [193] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [199] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [205] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [211] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [217] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [223] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [229] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [235] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [241] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [247] 1.00000000 1.00000000 1.00000000 0.02221093 1.00000000 1.00000000
#> [253] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [259] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [265] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [271] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [277] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [283] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [289] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [295] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [301] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [307] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [313] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [319] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [325] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [331] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [337] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [343] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [349] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [355] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [361] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [367] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [373] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [379] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [385] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [391] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [397] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [403] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [409] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [415] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [421] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [427] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [433] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [439] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [445] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [451] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [457] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [463] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [469] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [475] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [481] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [487] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [493] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [499] 1.00000000 1.00000000
#>  [ reached getOption("max.print") -- omitted 859 entries ]
get_pval(x, "chrM.1000", "bm", "fdr")
#>   [1] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>   [6] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [11] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [16] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [21] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [26] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [31] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [36] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [41] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [46] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [51] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [56] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [61] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [66] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [71] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [76] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [81] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [86] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [91] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [96] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [101] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [106] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [111] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [116] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [121] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [126] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [131] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [136] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [141] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [146] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [151] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [156] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [161] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [166] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [171] 1.682361e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [176] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [181] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [186] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [191] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [196] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 9.136479e-01
#> [201] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [206] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [211] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [216] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [221] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [226] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [231] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [236] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [241] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [246] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.675306e-06
#> [251] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [256] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [261] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [266] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [271] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [276] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [281] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [286] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [291] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [296] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [301] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [306] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [311] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [316] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [321] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [326] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [331] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [336] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [341] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [346] 1.000000e+00 1.000000e+00 1.000000e+00 9.136479e-01 1.000000e+00
#> [351] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [356] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [361] 1.000000e+00 1.000000e+00 1.000000e+00 9.136479e-01 1.000000e+00
#> [366] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [371] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [376] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [381] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [386] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [391] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [396] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [401] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [406] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [411] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [416] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [421] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [426] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [431] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [436] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [441] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [446] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [451] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [456] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [461] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [466] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [471] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [476] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [481] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [486] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [491] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [496] 9.136479e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [ reached getOption("max.print") -- omitted 859 entries ]
get_pval(x, "chrM.1000", "bi", "fdr")
#>   [1] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>   [6] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [11] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [16] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [21] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [26] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [31] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [36] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [41] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [46] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [51] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [56] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [61] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [66] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [71] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [76] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [81] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [86] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [91] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [96] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [101] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [106] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [111] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [116] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [121] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [126] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [131] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [136] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [141] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [146] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [151] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [156] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [161] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [166] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [171] 7.764643e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [176] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [181] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [186] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [191] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [196] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [201] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [206] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [211] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [216] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [221] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [226] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [231] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [236] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [241] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [246] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 3.765983e-05
#> [251] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [256] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [261] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [266] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [271] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [276] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [281] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [286] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [291] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [296] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [301] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [306] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [311] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [316] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [321] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [326] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [331] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [336] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [341] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [346] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [351] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [356] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [361] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [366] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [371] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [376] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [381] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [386] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [391] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [396] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [401] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [406] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [411] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [416] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [421] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [426] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [431] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [436] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [441] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [446] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [451] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [456] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [461] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [466] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [471] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [476] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [481] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [486] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [491] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [496] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [ reached getOption("max.print") -- omitted 859 entries ]