Open H5 file

This function opens the H5 file and create a mtmutObj object.

open_h5_file(h5_file)

Arguments

h5_file

a string of the h5 file directory

Value

a mtmutObj object

Details

The mtmutObj object is a S3 class for handling mitochondrial mutation data. It contains the following elements:

file

a string of the h5 file directory.

h5f

H5 file handle.

mut_table

allele count table H5 group handle.

loc_list

list of available loci.

loc_selected

selected loci, the default is all loci.

cell_list

list of available cell ids.

cell_selected

selected cell ids, the default is all cells.

loc_pass

loci passed the filter, the default is NULL

loc_filter

filter parameters.

loc_filter$min_cell

a integer of the minimum number of cells with mutation, the default is 1.

loc_filter$model

a string of the model for mutation calling, it can be "bb", "bm" or "bi", the default is "bb".

loc_filter$p_threshold

a numeric of the p-value threshold, the default is 0.05.

loc_filter$p_adj_method

a string of the method for p-value adjustment, refer to p.adjust, the default is "fdr".

Examples

## Use the example data
f <- system.file("extdata", "mini_dataset.tsv.gz", package = "scMitoMut")
## Create a temporary h5 file
## In real case, we keep the h5 in project folder for future use
f_h5_tmp <- tempfile(fileext = ".h5")
## Load the data with parse_table function
f_h5 <- parse_table(f, sep = "\t", h5_file = f_h5_tmp)
f_h5
#> [1] "/var/folders/46/vh_b6qzs5kzgdhvx1kcn7hyh0000gn/T//RtmpvCtNq4/file1555436a48bc.h5"
## open the h5 file and create a mtmutObj object
x <- open_h5_file(f_h5)
x
#> mtmutObj object
#> -------------------------------------------------
#> h5 file: /var/folders/46/vh_b6qzs5kzgdhvx1kcn7hyh0000gn/T//RtmpvCtNq4/file1555436a48bc.h5
#> Available loci: 16
#> Selected loci: 16
#> Available cells: 1359
#> Selected cells: 1359
#> Loci passed the filter: 0
#> filter parameters: 
#> 	min_cell: 1
#> 	model: bb
#> 	p_threshold: 0.05
#> 	p_adj_method: fdr