Load allele count table

This function loads the allele count table and save it to a H5 file.

parse_table(file, h5_file = "mut.h5", ...)

Arguments

file

a string of the allele count table file directory.

h5_file

a string of the output h5 file directory.

...

other parameters passed to fread.

Value

a string of the output h5 file directory.

Details

The allele count table should be a data.table with the following columns:

loc

a string of the locus

cell_barcode

a string of the cell barcode.

fwd_depth

a integer of the forward depth.

rev_depth

a integer of the reverse depth.

alt

a string of the alternative base.

ref

a string of the reference base.

coverage

a integer of the coverage.

Examples

## Use the example data
f <- system.file("extdata", "mini_dataset.tsv.gz", package = "scMitoMut")

## Create a temporary h5 file
## In real case, we keep the h5 in project folder for future use
f_h5_tmp <- tempfile(fileext = ".h5")

## Load the data with parse_table function
f_h5 <- parse_table(f, sep = "\t", h5_file = f_h5_tmp)
f_h5
#> [1] "/var/folders/46/vh_b6qzs5kzgdhvx1kcn7hyh0000gn/T//RtmpvCtNq4/file1555292a4c86.h5"

## open the h5 file and create a mtmutObj object
x <- open_h5_file(f_h5)
x
#> mtmutObj object
#> -------------------------------------------------
#> h5 file: /var/folders/46/vh_b6qzs5kzgdhvx1kcn7hyh0000gn/T//RtmpvCtNq4/file1555292a4c86.h5
#> Available loci: 16
#> Selected loci: 16
#> Available cells: 1359
#> Selected cells: 1359
#> Loci passed the filter: 0
#> filter parameters: 
#> 	min_cell: 1
#> 	model: bb
#> 	p_threshold: 0.05
#> 	p_adj_method: fdr