Fit tree models for one locus

This function fit binomial mixture model, beta binomial model and calculate the VMR and consistency of fwd rev strand for one locus.

process_locus_bmbb(mtmutObj, loc, dom_allele = NULL, return_data = FALSE, ...)

Arguments

mtmutObj

a mtmutObj object.

loc

string given the locus name (e.g. "chrM1000").

dom_allele

string given the dominant allele (e.g. "A"), if NULL auto detect the dominant allele.

return_data

logical whether to return the allele count data, if FALSE, the data in the return value will be NULL. The default is FALSE.

...

other parameters control the model fitting.

Value

A list of three elements:

data

data.frame of the allele count data.

locus

data.table of the VMR and consistency of fwd rev strand.

model

list of the model fitting results.

Examples

## Use the example data
f <- system.file("extdata", "mini_dataset.tsv.gz", package = "scMitoMut")

## Create a temporary h5 file
## In real case, we keep the h5 in project folder for future use
f_h5_tmp <- tempfile(fileext = ".h5")

## Load the data with parse_table function
f_h5 <- parse_table(f, sep = "\t", h5_file = f_h5_tmp)

x <- open_h5_file(f_h5)
res <- process_locus_bmbb(x, loc = "chrM.1000")
res
#> $data
#> NULL
#> 
#> $locus
#> $locus$mean_af
#> [1] 0.9991425
#> 
#> $locus$cor
#> [1] 0.9662561
#> 
#> $locus$vmr
#> [1] 2.931195e-05
#> 
#> 
#> $model
#> $model$beta_binom
#> $model$beta_binom$bb_pval
#>   [1] 1.000000e+00 2.621804e-02 1.000000e+00 1.000000e+00 1.000000e+00
#>   [6] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [11] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [16] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [21] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [26] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [31] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [36] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [41] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [46] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [51] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [56] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [61] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [66] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [71] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [76] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [81] 1.000000e+00 1.000000e+00 2.740668e-02 1.000000e+00 1.000000e+00
#>  [86] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [91] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [96] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [101] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [106] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [111] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [116] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [121] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [126] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [131] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [136] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [141] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [146] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [151] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 2.799714e-02
#> [156] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [161] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [166] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [171] 1.213369e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [176] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [181] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [186] 4.091438e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [191] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [196] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.822987e-02
#> [201] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [206] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [211] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [216] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [221] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [226] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [231] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [236] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [241] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [246] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.634358e-05
#> [251] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [256] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [261] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [266] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [271] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [276] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [281] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [286] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [291] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [296] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [301] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [306] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [311] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [316] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [321] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [326] 1.000000e+00 1.000000e+00 1.000000e+00 7.394679e-02 1.000000e+00
#> [331] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [336] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [341] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [346] 1.000000e+00 1.000000e+00 1.000000e+00 1.886110e-02 1.000000e+00
#> [351] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [356] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [361] 1.000000e+00 1.000000e+00 1.000000e+00 1.948945e-02 1.000000e+00
#> [366] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [371] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [376] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [381] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [386] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [391] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [396] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [401] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [406] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [411] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [416] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [421] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [426] 1.000000e+00 4.144895e-02 1.000000e+00 1.000000e+00 1.000000e+00
#> [431] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [436] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [441] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [446] 1.000000e+00 1.000000e+00 6.884048e-02 1.000000e+00 1.000000e+00
#> [451] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [456] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [461] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [466] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [471] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [476] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [481] 1.000000e+00 1.000000e+00 7.184342e-02 1.000000e+00 1.000000e+00
#> [486] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [491] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [496] 1.759573e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [ reached getOption("max.print") -- omitted 859 entries ]
#> 
#> $model$beta_binom$parameters
#>      mean_af      prob    theta
#> 1: 0.9991425 0.9992806 227.0007
#> 
#> 
#> $model$binom_mix
#> $model$binom_mix$bm_pval
#>   [1] 1.000000e+00 1.558707e-02 1.000000e+00 1.000000e+00 1.000000e+00
#>   [6] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [11] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [16] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [21] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [26] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [31] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [36] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [41] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [46] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [51] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [56] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [61] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [66] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [71] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [76] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [81] 1.000000e+00 1.000000e+00 1.636001e-02 1.000000e+00 1.000000e+00
#>  [86] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [91] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [96] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [101] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [106] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [111] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [116] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [121] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [126] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [131] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [136] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [141] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [146] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [151] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.674626e-02
#> [156] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [161] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [166] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [171] 8.665583e-04 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [176] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [181] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [186] 2.558814e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [191] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [196] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.054811e-02
#> [201] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [206] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [211] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [216] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [221] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [226] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [231] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [236] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [241] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [246] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.232749e-09
#> [251] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [256] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [261] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [266] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [271] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [276] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [281] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [286] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [291] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [296] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [301] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [306] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [311] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [316] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [321] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [326] 1.000000e+00 1.000000e+00 1.000000e+00 5.201197e-02 1.000000e+00
#> [331] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [336] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [341] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [346] 1.000000e+00 1.000000e+00 1.000000e+00 1.093663e-02 1.000000e+00
#> [351] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [356] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [361] 1.000000e+00 1.000000e+00 1.000000e+00 1.132501e-02 1.000000e+00
#> [366] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [371] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [376] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [381] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [386] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [391] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [396] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [401] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [406] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [411] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [416] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [421] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [426] 1.000000e+00 2.597076e-02 1.000000e+00 1.000000e+00 1.000000e+00
#> [431] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [436] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [441] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [446] 1.000000e+00 1.000000e+00 4.753361e-02 1.000000e+00 1.000000e+00
#> [451] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [456] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [461] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [466] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [471] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [476] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [481] 1.000000e+00 1.000000e+00 5.014855e-02 1.000000e+00 1.000000e+00
#> [486] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [491] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [496] 1.015943e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [ reached getOption("max.print") -- omitted 859 entries ]
#> 
#> $model$binom_mix$bi_pval
#>   [1] 1.000000e+00 3.373183e-02 1.000000e+00 1.000000e+00 1.000000e+00
#>   [6] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [11] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [16] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [21] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [26] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [31] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [36] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [41] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [46] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [51] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [56] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [61] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [66] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [71] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [76] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [81] 1.000000e+00 1.000000e+00 3.538823e-02 1.000000e+00 1.000000e+00
#>  [86] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [91] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [96] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [101] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [106] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [111] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [116] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [121] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [126] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [131] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [136] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [141] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [146] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [151] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 3.621536e-02
#> [156] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [161] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [166] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [171] 3.999448e-03 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [176] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [181] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [186] 5.504493e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [191] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [196] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 2.289569e-02
#> [201] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [206] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [211] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [216] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [221] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [226] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [231] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [236] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [241] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [246] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 2.771143e-08
#> [251] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [256] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [261] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [266] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [271] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [276] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [281] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [286] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [291] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [296] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [301] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [306] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [311] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [316] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [321] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [326] 1.000000e+00 1.000000e+00 1.000000e+00 1.101187e-01 1.000000e+00
#> [331] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [336] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [341] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [346] 1.000000e+00 1.000000e+00 1.000000e+00 2.373353e-02 1.000000e+00
#> [351] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [356] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [361] 1.000000e+00 1.000000e+00 1.000000e+00 2.457066e-02 1.000000e+00
#> [366] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [371] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [376] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [381] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [386] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [391] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [396] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [401] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [406] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [411] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [416] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [421] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [426] 1.000000e+00 5.585520e-02 1.000000e+00 1.000000e+00 1.000000e+00
#> [431] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [436] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [441] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [446] 1.000000e+00 1.000000e+00 1.009108e-01 1.000000e+00 1.000000e+00
#> [451] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [456] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [461] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [466] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [471] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [476] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [481] 1.000000e+00 1.000000e+00 1.062936e-01 1.000000e+00 1.000000e+00
#> [486] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [491] 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> [496] 2.205712e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#>  [ reached getOption("max.print") -- omitted 859 entries ]
#> 
#> $model$binom_mix$parameters
#>    loglik_k1 loglik_k2     k1_pi    k2_pi1    k2_pi2 k2_theta1
#> 1: -180.7086 -165.8746 0.9991425 0.9996073 0.9828958  0.972772
#> 
#> 
#>